+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.0
+- Add an additional check to make sure the decimal seperator and the group seperator are different.
+- Remove the flags from `patchNumberSeperators` that would allow for seperate patching of the group and decimal seperator.
+- remove the `shouldPatchLocales` flag from the `patchNumberSeperators`. 
+
 ## 1.2.0
 
 - Add the `isUsingSamsungKeyboard` function to detect if the user is using a samsung keyboard